refactor: move openapi configuration to autoconfigure
