+++ ff4j-spring-boot-autoconfigure/src/test/java/org/ff4j/spring/boot/autoconfigure/FF4JAuditEnableConfigurationTest.java
@@ -1,4 +1,4 @@
-package org.ff4j.spring.boot.web.api.config;
+package org.ff4j.spring.boot.autoconfigure;
 
 /*-
  * #%L
@@ -9,9 +9,9 @@ package org.ff4j.spring.boot.web.api.config;
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -23,8 +23,6 @@ package org.ff4j.spring.boot.web.api.config;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import org.ff4j.FF4j;
-import org.ff4j.spring.boot.autoconfigure.FF4JConfiguration;
-import org.ff4j.spring.boot.web.api.Application;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
